Move InvokeForTableRequest to core module
